AquaCrop

Documentation for AquaCrop.

AquaCrop.AbstractParametersContainer

AquaCrop.CompartmentIndividual

AquaCrop.FieldAquaCrop

AquaCrop.FieldAquaCrop

AquaCrop.ParametersContainer

AquaCrop.ProjectInputType

AquaCrop.RepAssimilates

AquaCrop.RepClim

AquaCrop.RepContent

AquaCrop.RepCrop

AquaCrop.RepCropFileSet

AquaCrop.RepCuttings

AquaCrop.RepDayEventDbl

AquaCrop.RepDayEventInt

AquaCrop.RepEffectStress

AquaCrop.RepEffectiveRain

AquaCrop.RepEndSeason

AquaCrop.RepFileOK

AquaCrop.RepIniSWC

AquaCrop.RepIrriECw

AquaCrop.RepManag

AquaCrop.RepOnset

AquaCrop.RepParam

AquaCrop.RepPerennialPeriod

AquaCrop.RepShapes

AquaCrop.RepSim

AquaCrop.RepSoil

AquaCrop.RepStorage

AquaCrop.RepSum

AquaCrop.SoilLayerIndividual

AquaCrop.active_cells

AquaCrop.adjust_calendar_crop!

AquaCrop.adjust_calendar_days!

AquaCrop.adjust_climrecord_to!

AquaCrop.adjust_compartments!

AquaCrop.adjust_crop_file_parameters!

AquaCrop.adjust_crop_year_to_climfile!

AquaCrop.adjust_decade_month_and_year

AquaCrop.adjust_month_and_year

AquaCrop.adjust_onset_search_period!

AquaCrop.adjust_simperiod!

AquaCrop.adjust_size_compartments!

AquaCrop.adjust_theta_initial!

AquaCrop.adjust_year_perennials!

AquaCrop.calculate_adjusted_fc!

AquaCrop.calculate_salt_mobility!

AquaCrop.check_files_in_project!

AquaCrop.check_for_keep_swc

AquaCrop.complete_climate_description!

AquaCrop.complete_crop_description!

AquaCrop.complete_profile_description!

AquaCrop.crop_stress_parameters_soil_fertility!

AquaCrop.dailyupdate!

AquaCrop.days_to_reach_cc_with_given_cgc

AquaCrop.default_filepaths

AquaCrop.degrees_day

AquaCrop.derive_smax_top_bottom

AquaCrop.designate_soillayer_to_compartments!

AquaCrop.determinate_coeffcapillaryrise!

AquaCrop.determinate_soilclass!

AquaCrop.determine_date

AquaCrop.determine_day_nr

AquaCrop.determine_length_growth_stages

AquaCrop.determine_linked_simday1!

AquaCrop.determine_nrand_thickness_compartments!

AquaCrop.determine_salt_content!

AquaCrop.ececomp

AquaCrop.find_values

AquaCrop.from_gravelmass_to_gravelvol

AquaCrop.full_undefined_record

AquaCrop.gddcdc_to_cdc

AquaCrop.get_decade_temperature_dataset!

AquaCrop.get_interpolation_parameters

AquaCrop.get_monthly_temperature_dataset!

AquaCrop.get_parameters

AquaCrop.get_project_type

AquaCrop.get_results_parameters

AquaCrop.get_set_of_three

AquaCrop.get_set_of_three_months

AquaCrop.getparameter

AquaCrop.growing_degree_days

AquaCrop.initialize_project

AquaCrop.initialize_project_filename

AquaCrop.initialize_project_input

AquaCrop.initialize_run_part_1!

AquaCrop.initialize_settings

AquaCrop.initialize_the_program

AquaCrop.isleapyear

AquaCrop.ks_any

AquaCrop.length_canopy_decline

AquaCrop.load_clim!

AquaCrop.load_crop!

AquaCrop.load_groundwater!

AquaCrop.load_initial_conditions!

AquaCrop.load_irri_schedule_info!

AquaCrop.load_management!

AquaCrop.load_offseason!

AquaCrop.load_profile

AquaCrop.load_profile_processing!

AquaCrop.load_program_parameters_project_plugin!

AquaCrop.load_simulation_project!

AquaCrop.no_adjustment

AquaCrop.no_irrigation!

AquaCrop.read_month

AquaCrop.read_temperature_file!

AquaCrop.reset_previous_sum!

AquaCrop.reset_swc_to_fc!

AquaCrop.root_max_in_soil_profile

AquaCrop.run_simulation!

AquaCrop.salt_solution_deposit!

AquaCrop.set_clim_data!

AquaCrop.set_daynr_to_yundef

AquaCrop.setparameter!

AquaCrop.start_the_program

AquaCrop.sum_calendar_days

AquaCrop.tau_from_ksat

AquaCrop.time_to_cc_ini

AquaCrop.time_to_max_canopy_sf

AquaCrop.translate_inilayers_to_swprofile!

AquaCrop.translate_inipoints_to_swprofile!

AquaCrop.zr_adjusted_to_restrictive_layers

AquaCrop.AbstractParametersContainer — Type

AbstractParametersContainer

AquaCrop.CompartmentIndividual — Type

compartment = CompartmentIndividual()

AquaCrop.FieldAquaCrop — Type

FieldAquaCrop

Has all the data for the simulation of AquaCrop

AquaCrop.FieldAquaCrop — Method

FieldAquaCrop(inputDataDir::String)

Starts the struct FieldAquaCrop that has all the data for the simulation of AquaCrop

AquaCrop.ParametersContainer — Type

pc = ParametersContainer(T)

Contains parameters in a Dictionary of type Symbol=>T. it is used for parameters without a type in the original AquaCrop.f90

AquaCrop.ProjectInputType — Type

projectinput = ProjectInputType()

Container for project file input data.

AquaCrop.RepAssimilates — Type

assimilates = RepAssimilates()

set from defaultcropsoil.f90:274

AquaCrop.RepClim — Type

record_clim = RepClim()

AquaCrop.RepContent — Type

content = RepContent()

AquaCrop.RepCrop — Type

crop = RepCrop()

set from defaultcropsoil.f90:140

AquaCrop.RepCropFileSet — Type

crop_file_set = RepCropFileSet()

AquaCrop.RepCuttings — Type

cuttings = RepCuttings()

set from global.f90:3338

AquaCrop.RepDayEventDbl — Type

a = RepDayEventDbl()

AquaCrop.RepDayEventInt — Type

a = RepDayEventInt()

set from global.f90:2838

AquaCrop.RepEffectStress — Type

effectstress = RepEffectStress()

AquaCrop.RepEffectiveRain — Type

effectiverain = RepEffectiveRain()

set from initialsettings.f90:244

AquaCrop.RepEndSeason — Type

endseason = RepEndSeason()

AquaCrop.RepFileOK — Type

fileok = RepFileOK()

AquaCrop.RepIniSWC — Type

iniswc = RepIniSWC()

AquaCrop.RepIrriECw — Type

irriecw = RepIrriECw()

set from global.f90:2838

AquaCrop.RepManag — Type

management = RepManag()

set from global.f90:3311

AquaCrop.RepOnset — Type

onset = RepOnset()

set from initialsettings.f90:464

AquaCrop.RepParam — Type

simulparams = RepParams()

contains the simulation parameters.

set from initialsettings.f90:216

AquaCrop.RepPerennialPeriod — Type

perennial_period = RepPerennialPeriod()

AquaCrop.RepShapes — Type

shape = RepShapes()

set from defaultcropsoil.f90:175

AquaCrop.RepSim — Type

simulation = RepSim()

set from global.f90:7694

AquaCrop.RepSoil — Type

soil = RepSoil()

set from defaultcropsoil.f90:289

AquaCrop.RepStorage — Type

storage = RepStorage()

AquaCrop.RepSum — Type

summ = RepSum()

set from global.f90:7153

AquaCrop.SoilLayerIndividual — Type

soillayer = SoilLayerIndividual()

creates a soil layer with a given soil class.

set from defaultcropsoil.f90:289

AquaCrop.active_cells — Method

celi = active_cells(compartment::CompartmentIndividual, soil_layers::Vector{SoilLayerIndividual})

globa.f90:4241

AquaCrop.adjust_calendar_crop! — Method

adjust_calendar_crop!(inse)

tempprocessing.f90:1467

AquaCrop.adjust_calendar_days! — Method

adjust_calendar_days!(inse, iscgcgiven)

tempprocessing.f90:1327

AquaCrop.adjust_climrecord_to! — Method

adjust_climrecord_to!(clim_record::RepClim, cdayn)

global.f90:6163

AquaCrop.adjust_compartments! — Method

adjust_compartments!(inse)

run.f90:6619

AquaCrop.adjust_crop_file_parameters! — Method

adjust_crop_file_parameters!(inse)

tempprocessing.f90:1882

AquaCrop.adjust_crop_year_to_climfile! — Method

adjust_crop_year_to_climfile!(crop::RepCrop, clim_file, clim_record)

global.f90:6447

AquaCrop.adjust_decade_month_and_year — Method

decfile, mfile, yfile = adjust_decade_month_and_year(decfile, mfile, yfile)

tempprocessing.f90:293

AquaCrop.adjust_month_and_year — Method

mfile, yfile = adjust_month_and_year(mfile, yfile)

tempprocessing.f90:284

AquaCrop.adjust_onset_search_period! — Method

adjust_onset_search_period!(inse)

global.f90:4214

AquaCrop.adjust_simperiod! — Method

adjust_simperiod!(inse, projectinput::ProjectInputType)

global.f90:4692

AquaCrop.adjust_size_compartments! — Method

adjust_size_compartments!(inse, cropzx)

global.f90:6563

AquaCrop.adjust_theta_initial! — Method

adjust_theta_initial!(inse, prevnrcomp, prevthickcomp, prevvolprcomp, prevecdscomp)

global.f90:5852

AquaCrop.adjust_year_perennials! — Method

adjust_year_perennials!(crop::RepCrop, theyearseason)

global.f90:4690

AquaCrop.calculate_adjusted_fc! — Method

calculate_adjusted_fc!(compartadj::Vector{CompartmentIndividual}, soil_layers::Vector{SoilLayerIndividual}, depthaquifer)

global.f90:4141

AquaCrop.calculate_salt_mobility! — Method

calculate_salt_mobility!(soillayer::SoilLayerIndividual, saltdiffusion)

sets the salt mobility

global.f90:7500

AquaCrop.check_files_in_project! — Method

check_files_in_project!(fileok::RepFileOK, allok, input::ProjectInputType)

global.f90:7195

AquaCrop.check_for_keep_swc — Method

runwithkeepswc, constzrxforrun = checkforkeep_swc(projectinput::Vector{ProjectInputType}, filepaths, inse)

global.f90:6643

AquaCrop.complete_climate_description! — Method

complete_climate_description!(record::RepClim)

global.f90:8223

AquaCrop.complete_crop_description! — Method

complete_crop_description!(crop::RepCrop, simulation::RepSim, management::RepManag)

global.f90:5624

AquaCrop.complete_profile_description! — Method

complete_profile_description!(soil_layers::Vector{SoilLayerIndividual}, 
        compartments::Vector{CompartmentIndividual}, simulation::RepSim, total_water_content::RepContent)

global.f90:7561

AquaCrop.crop_stress_parameters_soil_fertility! — Method

crop_stress_parameters_soil_fertility!(stressout::RepEffectStress, cropsresp::RepShapes, stresslevel)

global.f90:1231

AquaCrop.dailyupdate! — Method

dailyupdate!(fieldcrop::AbstractFieldCrop)

Updates the FieldCrop by one day

AquaCrop.days_to_reach_cc_with_given_cgc — Method

daystoresult = days_to_reach_cc_with_given_cgc(cctoreach, ccoval, ccxval, cgcval, l0)

global.f90:1809

AquaCrop.default_filepaths — Method

fl = default_filepaths(parentdir::AbstractString)

sets the default directories for the input files.

startunit.f90:420

AquaCrop.degrees_day — Method

dgrd = degrees_day(tbase, tupper, tdaymin, tdaymax, gddselectedmethod)

global.f90:2419

AquaCrop.derive_smax_top_bottom — Method

sxtop, sxbot = derive_smax_top_bottom(crop::RepCrop)

global.f90:1944

AquaCrop.designate_soillayer_to_compartments! — Method

designate_soillayer_to_compartments!(compartments::Vector{CompartmentIndividual}, soil_layers::Vector{SoilLayerIndividual})

global.f90:7393

AquaCrop.determinate_coeffcapillaryrise! — Method

determinate_coeffcapillaryrise!(soillayer::SoilLayerIndividual)

sets the coefficients for capillary rise of soillayer considering its own data.

global.f90:4034

AquaCrop.determinate_soilclass! — Method

determinate_soilclass!(soillayer::SoilLayerIndividual)

sets the soil class of soillayer considering its own data.

global.f90:1909

AquaCrop.determine_date — Method

dayi, monthi, yeari = determine_date(dar_nr)

global.f90:2397

AquaCrop.determine_day_nr — Method

day_nr = determine_day_nr(dayi, monthi, yeari)

global.f90:2387

AquaCrop.determine_length_growth_stages — Method

length123, stlength, length12, cgcval = determine_length_growth_stages(ccoval, ccxval, cdcval, l0, totallength, 
                                                                    cgcgiven, thedaystoccini, theplanting, 
                                                                    length123, stlength, length12, cgcval)

global.f90:1644

AquaCrop.determine_linked_simday1! — Method

determine_linked_simday1!(simulation::RepSim, crop::RepCrop, clim_record::RepClim, clim_file)

global.f90:4677

AquaCrop.determine_nrand_thickness_compartments! — Method

determine_nrand_thickness_compartments!(compartments::Vector{CompartmentIndividual}, soil_layers::Vector{SoilLayerIndividual}, compdefthick)

global.f90:4063

AquaCrop.determine_salt_content! — Method

determine_salt_content!(compartment::CompartmentIndividual, soil_layers::Vector{SoilLayerIndividual}, simulparam::RepParam)

global.f90:4258

AquaCrop.ececomp — Method

ece = ececomp(compartment::CompartmentIndividual, inse)

global.f90:2523

AquaCrop.find_values — Method

zcn, ecdsm = find_values(atdaynr, daynr1, daynr2, z1, ec1, z2, ec2)

global.f90:6140

AquaCrop.from_gravelmass_to_gravelvol — Method

gravelvol = from_gravelmass_to_gravelvol(porositypercent, gravelmasspercent)

calculates the gravel volume of soil layer.

global.f90:1521

AquaCrop.full_undefined_record — Method

logi = full_undefined_record(record::RepClim)

global.f90:2826

AquaCrop.gddcdc_to_cdc — Method

cdc = gddcdc_to_cdc(plantdaynr, d123, gddl123, gddharvest, ccx, gddcdc, tbase, tupper, notempfiletmin, notempfiletmax, inse)

tempprocessing.f90:1545

AquaCrop.get_decade_temperature_dataset! — Method

get_decade_temperature_dataset!(tmin_dataset, tmax_dataset, daynri, temperature_file, temperature_record::RepClim)

tempprocessing.f90:362

AquaCrop.get_interpolation_parameters — Method

 aover3, bover2, c = get_interpolation_parameters(c1, c2, c3)

tempprocessing.f90:854

AquaCrop.get_monthly_temperature_dataset! — Method

get_monthly_temperature_dataset!(tmin_dataset, tmax_dataset, daynri, temperature_file, temperature_record::RepClim)

tempprocessing.f90:596

AquaCrop.get_parameters — Method

ul, ll, mid = get_parameters(c1, c2, c3)

tempprocessing.f90:580

AquaCrop.get_project_type — Method

pt = get_project_type(theprojectfile)

gets the project type for a given file.

startunit.f90:322

AquaCrop.get_results_parameters — Method

resultsparameters = get_results_parameters(path::String)

gets all the results parameters in filepaths[:simul].

startunit.f90:426

AquaCrop.get_set_of_three — Method

c1min, c1max, c2min, c2max, c3min, c3max = get_set_of_three(dayn, deci, monthi, yeari, temperature_file, temperature_record::RepClim)

tempprocessing.f90:439

AquaCrop.get_set_of_three_months — Method

c1min, c2min, c3min, c1max, c2max, c3max, x1, x2, x3, t1 = get_set_of_three_months(monthi, yeari, temperature_file, temperature_record::RepClim)

tempprocessing.f90:645

AquaCrop.getparameter — Method

getparameter(parameterscontainer::ParametersContainer{T}, parameterkey::Symbol)

gets the parameterkey from parameterscontainer if parameterkey does not exist returns missing.

AquaCrop.growing_degree_days — Method

gdd1234 = growing_degree_days(inse, tdaymin, tdaymax)

tempprocessing.f90:871

AquaCrop.initialize_project — Method

inse, projectinput, fileok = initialize_project(i, theprojectfile, theprojecttype, filepaths)

startunit.f90:535

AquaCrop.initialize_project_filename — Method

project_filenames = initialize_project_filename(filepaths)

Gets all the names of the projects files.

startunit.f90:441

AquaCrop.initialize_project_input — Method

projectinput = initialize_project_input(filename, parentdir)

project_input.f90:152

AquaCrop.initialize_run_part_1! — Method

initialize_run_part_1!(inse, projectinput::ProjectInputType)

run.f90:6590

AquaCrop.initialize_settings — Method

inse = initialize_settings(usedefaultsoilfile, usedefaultcropfile, filepaths)

gets the initial settings.

initialsettings.f90:201

AquaCrop.initialize_the_program — Method

filepaths, resultsparameters = initialize_the_program(parentdir)

Gets the file paths and the simulation parameters.

startunit.f90:417

AquaCrop.isleapyear — Method

leapyear = isleapyear(year)

global.f90:7100

AquaCrop.ks_any — Method

ksval = ks_any(wrel, pulactual, pllactual, shapefactor)

global.f90:2611

AquaCrop.length_canopy_decline — Method

nd = length_canopy_decline(ccx, cdc)

global.f90:1839

AquaCrop.load_clim! — Method

load_clim!(record::RepClim, filename)

global.f90:5936

AquaCrop.load_crop! — Method

load_crop!(crop::RepCrop, perennial_period::RepPerennialPeriod, crop_file)

global.f90:4799

AquaCrop.load_groundwater! — Method

load_groundwater!(inse, fullname)

global.f90:5981

AquaCrop.load_initial_conditions! — Method

 load_initial_conditions!(inse, swcinifilefull)

global.f90:6486

AquaCrop.load_irri_schedule_info! — Method

load_irri_schedule_info!(inse, fullname)

global.f90:2860

AquaCrop.load_management! — Method

load_management!(inse, fullname)

global.f90:3350

AquaCrop.load_offseason! — Method

load_offseason!(inse, fullname)

global.f90:5769

AquaCrop.load_profile — Method

soil, soil_layers, compartments = load_profile(filepath, simulparam::RepParam)

loads data from filepath.

global.f90:7590

AquaCrop.load_profile_processing! — Method

load_profile_processing!(soil::RepSoil, soil_layers::Vector{SoilLayerIndividual},
    compartments::Vector{CompartmentIndividual}, simulparam::RepParam)

loads some data.

global.f90:7684

AquaCrop.load_program_parameters_project_plugin! — Method

load_program_parameters_project_plugin!(simulparam::RepParam, auxparfile)

startunit.f90:767

AquaCrop.load_simulation_project! — Method

load_simulation_project!(inse, projectinput::ProjectInputType)

tempprocessing.f90:1932

AquaCrop.no_adjustment — Method

nadj = no_adjustment(fcvolpr)

global.f90:4195

AquaCrop.no_irrigation! — Method

no_irrigation!(inse)

global.f90:2838

AquaCrop.read_month — Method

cimin, cimax = read_month(stringline)

tempprocessing.f90:837

AquaCrop.read_temperature_file! — Method

read_temperature_file!(array_parameters::ParametersContainer{T}, temperature_file) where T

tempprocessing.f90:307

AquaCrop.reset_previous_sum! — Method

reset_previous_sum!(inse)

run.f90:3445

AquaCrop.reset_swc_to_fc! — Method

reset_swc_to_fc!(simulation::RepSim, compartments::Vector{CompartmentIndividual},
                 soil_layers::Vector{SoilLayerIndividual}, ziaqua)

global.f90:4758

AquaCrop.root_max_in_soil_profile — Method

zmax = root_max_in_soil_profile(zmaxcrop, soil_layers::Vector{SoilLayerIndividual})

global.f90:1092

AquaCrop.run_simulation! — Method

run_simulation!(inse, projectinput::Vector{ProjectInputType})

run.f90:7779

AquaCrop.salt_solution_deposit! — Method

salt_solution_deposit!(compartment::CompartmentIndividual, simulparam::RepParam, i, mm)

global.f90:2572

AquaCrop.set_clim_data! — Method

set_clim_data!(inse, projectinput::ProjectInputType)

global.f90:4295

AquaCrop.set_daynr_to_yundef — Method

daynri = set_daynr_to_yundef(daynri)

tempprocessing.f90:351

AquaCrop.setparameter! — Method

setparameter!(parameterscontainer::ParametersContainer{T}, parameterkey::Symbol, parameter::T)

sets an entry in parameterscontainer.

AquaCrop.start_the_program — Function

start_the_program(dir::Union{String,Nothing}=nothing)

starts the program

startunit.f90:931

AquaCrop.sum_calendar_days — Method

nrcdays = sum_calendar_days(valgddays, firstdaycrop, tbase, tupper, tdaymin, tdaymax, inse)

tempprocessing.f90:1035

AquaCrop.tau_from_ksat — Method

tau = tau_from_ksat(ksat)

global.f90:1889

AquaCrop.time_to_cc_ini — Method

elapsedtime = time_to_cc_ini(theplantingtype, thecropplantingdens, 
                      thesizeseedling, thesizeplant, thecropccx, thecropcgc)

global.f90:1754

AquaCrop.time_to_max_canopy_sf — Method

l12sf, redcgc, redccx, classsf = time_to_max_canopy_sf(cco, cgc, ccx, l0, l12, l123, ltoflor, lflor, determinantcrop, l12sf, redcgc, redccx, classsf)

global.f90:2090

AquaCrop.translate_inilayers_to_swprofile! — Method

translate_inilayers_to_swprofile!(inse, nrlay, laythickness, layvolpr, layecds)

global.f90:6179

AquaCrop.translate_inipoints_to_swprofile! — Method

translate_inipoints_to_swprofile!(inse, nrloc, locdepth, locvolpr, locecds)

global.f90:6258

AquaCrop.zr_adjusted_to_restrictive_layers — Method

zrout = zr_adjusted_to_restrictive_layers(zrin, soil_layers::Vector{SoilLayerIndividual})

global.f90:1126