AquaCrop

Documentation for AquaCrop.

AquaCrop.AbstractParametersContainer

AquaCrop.CompartmentIndividual

AquaCrop.FieldCrop

AquaCrop.FieldCrop

AquaCrop.ParametersContainer

AquaCrop.ProjectInputType

AquaCrop.RepAssimilates

AquaCrop.RepContent

AquaCrop.RepCrop

AquaCrop.RepCuttings

AquaCrop.RepDayEventInt

AquaCrop.RepEffectStress

AquaCrop.RepEffectiveRain

AquaCrop.RepEndSeason

AquaCrop.RepFileOK

AquaCrop.RepIniSWC

AquaCrop.RepIrriECw

AquaCrop.RepManag

AquaCrop.RepOnset

AquaCrop.RepParam

AquaCrop.RepShapes

AquaCrop.RepSim

AquaCrop.RepSoil

AquaCrop.RepStorage

AquaCrop.RepSum

AquaCrop.SoilLayerIndividual

AquaCrop.calculate_saltmobility!

AquaCrop.checkfilesinproject!

AquaCrop.checkforkeepswc

AquaCrop.completecropdescription!

AquaCrop.completeprofiledescription!

AquaCrop.cropstressparameterssoilfertility!

AquaCrop.dailyupdate!

AquaCrop.daystoreachccwithgivencgc

AquaCrop.defaultfilepaths

AquaCrop.designate_soillayer_to_compartments!

AquaCrop.determinate_coeffcapillaryrise!

AquaCrop.determinate_soilclass!

AquaCrop.determinelengthgrowthstages

AquaCrop.determinenrandthicknesscompartments!

AquaCrop.from_gravelmass_to_gravelvol

AquaCrop.getparameter

AquaCrop.getprojecttype

AquaCrop.getresultsparameters

AquaCrop.initializeproject

AquaCrop.initializeprojectfilename

AquaCrop.initializeprojectinput

AquaCrop.initializesettings

AquaCrop.initializetheprogram

AquaCrop.ksany

AquaCrop.lengthcanopydecline

AquaCrop.loadprofile

AquaCrop.loadprofileprocessing!

AquaCrop.loadprogramparametersprojectplugin!

AquaCrop.rootmaxinsoilprofile

AquaCrop.setparameter!

AquaCrop.taufromksat

AquaCrop.timetoccini

AquaCrop.timetomaxcanopysf

AquaCrop.zradjustedtorestrictivelayers

AquaCrop.AbstractParametersContainer — Type

AbstractParametersContainer

AquaCrop.CompartmentIndividual — Type

compartment = CompartmentIndividual()

AquaCrop.FieldCrop — Type

FieldCrop

Has all the data for the simulation of AquaCrop

AquaCrop.FieldCrop — Method

FieldCrop(inputDataDir::String)

Starts the struct FieldCrop that has all the data for the simulation of AquaCrop

AquaCrop.ParametersContainer — Type

pc = ParametersContainer(T)

Contains parameters in a Dictionary of type Symbol=>T. it is used for parameters without a type in the original AquaCrop.f90

AquaCrop.ProjectInputType — Type

projectinput = ProjectInputType()

Container for project file input data.

AquaCrop.RepAssimilates — Type

assimilates = RepAssimilates()

AquaCrop.RepContent — Type

content = RepContent()

AquaCrop.RepCrop — Type

crop = RepCrop()

AquaCrop.RepCuttings — Type

cuttings = RepCuttings()

AquaCrop.RepDayEventInt — Type

a = RepDayEventInt()

AquaCrop.RepEffectStress — Type

effectstress = RepEffectStress()

AquaCrop.RepEffectiveRain — Type

effectiverain = RepEffectiveRain()

AquaCrop.RepEndSeason — Type

endseason = RepEndSeason()

AquaCrop.RepFileOK — Type

fileok = RepFileOK()

AquaCrop.RepIniSWC — Type

iniswc = RepIniSWC()

AquaCrop.RepIrriECw — Type

irriecw = RepIrriECw()

AquaCrop.RepManag — Type

management = RepManag()

AquaCrop.RepOnset — Type

onset = RepOnset()

AquaCrop.RepParam — Type

simulparams = RepParams()

contains the simulation parameters.

AquaCrop.RepShapes — Type

shape = RepShapes()

AquaCrop.RepSim — Type

simulation = RepSim()

AquaCrop.RepSoil — Type

soil = RepSoil()

AquaCrop.RepStorage — Type

storage = RepStorage()

AquaCrop.RepSum — Type

summ = RepSum()

AquaCrop.SoilLayerIndividual — Type

soillayer = SoilLayerIndividual()

creates a soil layer with a given soil class.

AquaCrop.calculate_saltmobility! — Method

calculate_saltmobility!(soillayer::SoilLayerIndividual, saltdiffusion)

sets the salt mobility

AquaCrop.checkfilesinproject! — Method

checkfilesinproject!(fileok::RepFileOK, allok, input::ProjectInputType, parentdir)

AquaCrop.checkforkeepswc — Method

runwithkeepswc, constzrxforrun = checkforkeepswc(projectinput::Vector{ProjectInputType}, filepaths, inse)

AquaCrop.completecropdescription! — Method

completecropdescription!(crop::RepCrop, simulation::RepSim, management::RepManag)

AquaCrop.completeprofiledescription! — Method

completeprofiledescription!(soillayers::Vector{SoilLayerIndividual}, 
        compartments::Vector{CompartmentIndividual}, simulation::RepSim, totalwatercontent::RepContent)

AquaCrop.cropstressparameterssoilfertility! — Method

cropstressparameterssoilfertility!(stressout::RepEffectStress, cropsresp::RepShapes, stresslevel)

AquaCrop.dailyupdate! — Method

dailyupdate!(fieldcrop::AbstractFieldCrop)

Updates the FieldCrop by one day

AquaCrop.daystoreachccwithgivencgc — Method

daystoresult = daystoreachccwithgivencgc(cctoreach, ccoval, ccxval, cgcval, l0)

AquaCrop.defaultfilepaths — Method

fl = defaultfilepaths(parentdir::AbstractString)

sets the default directories for the input files.

AquaCrop.designate_soillayer_to_compartments! — Method

designate_soillayer_to_compartments!(compartments::Vector{CompartmentIndividual}, soillayers::Vector{SoilLayerIndividual})

AquaCrop.determinate_coeffcapillaryrise! — Method

determinate_coeffcapillaryrise!(soillayer::SoilLayerIndividual)

sets the coefficients for capillary rise of soillayer considering its own data.

AquaCrop.determinate_soilclass! — Method

determinate_soilclass!(soillayer::SoilLayerIndividual)

sets the soil class of soillayer considering its own data.

AquaCrop.determinelengthgrowthstages — Method

length123, stlength, length12, cgcval = determinelengthgrowthstages(ccoval, ccxval, cdcval, l0, totallength, 
                                                                    cgcgiven, thedaystoccini, theplanting, 
                                                                    length123, stlength, length12, cgcval)

AquaCrop.determinenrandthicknesscompartments! — Method

determinenrandthicknesscompartments!(compartments::Vector{CompartmentIndividual}, soillayers::Vector{SoilLayerIndividual}, compdefthick)

AquaCrop.from_gravelmass_to_gravelvol — Method

gravelvol = from_gravelmass_to_gravelvol(porositypercent, gravelmasspercent)

calculates the gravel volume of soil layer.

AquaCrop.getparameter — Method

getparameter(parameterscontainer::ParametersContainer{T}, parameterkey::Symbol)

gets the parameterkey from parameterscontainer if parameterkey does not exist returns missing.

AquaCrop.getprojecttype — Method

pt = getprojecttype(theprojectfile)

gets the project type for a given file.

AquaCrop.getresultsparameters — Method

resultsparameters = getresultsparameters(path::String)

gets all the results parameters in filepaths[:simul].

AquaCrop.initializeproject — Method

inse, projectinput, fileok = initializeproject(i, theprojectfile, theprojecttype, filepaths)

AquaCrop.initializeprojectfilename — Method

projectfilenames = initializeprojectfilename(filepaths)

Gets all the names of the projects files.

AquaCrop.initializeprojectinput — Method

projectinput = initializeprojectinput(filename)

AquaCrop.initializesettings — Method

inse = initializesettings(usedefaultsoilfile, usedefaultcropfile, filepaths)

gets the initial settings.

AquaCrop.initializetheprogram — Method

filepaths, resultsparameters = initializetheprogram(parentdir)

Gets the file paths and the simulation parameters.

AquaCrop.ksany — Method

ksval = ksany(wrel, pulactual, pllactual, shapefactor)

AquaCrop.lengthcanopydecline — Method

nd = lengthcanopydecline(ccx, cdc)

AquaCrop.loadprofile — Method

soil, soillayers, compartments = loadprofile(filepath, simulparam::RepParam)

loads data from filepath.

AquaCrop.loadprofileprocessing! — Method

loadprofileprocessing!(soil::RepSoil, soillayers::Vector{SoilLayerIndividual},
    compartments::Vector{CompartmentIndividual}, simulparam::RepParam)

loads some data.

AquaCrop.loadprogramparametersprojectplugin! — Method

loadprogramparametersprojectplugin!(simulparam::RepParam, auxparfile)

AquaCrop.rootmaxinsoilprofile — Method

zmax = rootmaxinsoilprofile(zmaxcrop, soillayers::Vector{SoilLayerIndividual})

AquaCrop.setparameter! — Method

setparameter!(parameterscontainer::ParametersContainer{T}, parameterkey::Symbol, parameter::T)

sets an entry in parameterscontainer.

AquaCrop.taufromksat — Method

tau = taufromksat(ksat)

AquaCrop.timetoccini — Method

elapsedtime = timetoccini(theplantingtype, thecropplantingdens, 
                      thesizeseedling, thesizeplant, thecropccx, thecropcgc)

AquaCrop.timetomaxcanopysf — Method

l12sf, redcgc, redccx, classsf = timetomaxcanopysf(cco, cgc, ccx, l0, l12, l123, ltoflor, lflor, determinantcrop, l12sf, redcgc, redccx, classsf)

AquaCrop.zradjustedtorestrictivelayers — Method

zrout = zradjustedtorestrictivelayers(zrin, soillayers::Vector{SoilLayerIndividual})