AquaCrop

Documentation for AquaCrop.

AquaCrop.ton

AquaCrop.AbstractParametersContainer

AquaCrop.CompartmentIndividual

AquaCrop.FieldAquaCrop

AquaCrop.FieldAquaCrop

AquaCrop.ParametersContainer

AquaCrop.ProjectInputType

AquaCrop.RepAssimilates

AquaCrop.RepClim

AquaCrop.RepContent

AquaCrop.RepCrop

AquaCrop.RepCropFileSet

AquaCrop.RepCutInfoRecord

AquaCrop.RepCuttings

AquaCrop.RepDayEventDbl

AquaCrop.RepDayEventInt

AquaCrop.RepEffectStress

AquaCrop.RepEffectiveRain

AquaCrop.RepEndSeason

AquaCrop.RepFileOK

AquaCrop.RepGwTable

AquaCrop.RepIniSWC

AquaCrop.RepIrriECw

AquaCrop.RepIrriInfoRecord

AquaCrop.RepManag

AquaCrop.RepOnset

AquaCrop.RepParam

AquaCrop.RepPerennialPeriod

AquaCrop.RepPlotPar

AquaCrop.RepRootZoneSalt

AquaCrop.RepRootZoneWC

AquaCrop.RepShapes

AquaCrop.RepSim

AquaCrop.RepSoil

AquaCrop.RepStorage

AquaCrop.RepStressTot

AquaCrop.RepSum

AquaCrop.RepTransfer

AquaCrop.SoilLayerIndividual

AquaCrop.StressIndexesBio

AquaCrop.StressIndexesSalt

AquaCrop.active_cells

AquaCrop.actual_rooting_depth

AquaCrop.actual_rooting_depth_days

AquaCrop.actual_rooting_depth_gddays

AquaCrop.actualize_with_dict!

AquaCrop.add_output_in_dayout!

AquaCrop.add_output_in_etodatasim!

AquaCrop.add_output_in_evaldataout!

AquaCrop.add_output_in_harvestsout!

AquaCrop.add_output_in_logger!

AquaCrop.add_output_in_raindatasim!

AquaCrop.add_output_in_seasonout!

AquaCrop.add_output_in_tcropsim!

AquaCrop.add_output_in_tempdatasim!

AquaCrop.adjust_calendar_crop!

AquaCrop.adjust_calendar_days!

AquaCrop.adjust_climrecord_to!

AquaCrop.adjust_compartments!

AquaCrop.adjust_crop_file_parameters!

AquaCrop.adjust_crop_year_to_climfile!

AquaCrop.adjust_decade_month_and_year

AquaCrop.adjust_epot_mulch_wetted_surface!

AquaCrop.adjust_for_watertable!

AquaCrop.adjust_month

AquaCrop.adjust_month_and_year

AquaCrop.adjust_onset_search_period!

AquaCrop.adjust_pleaf_to_eto

AquaCrop.adjust_psenescence_to_eto

AquaCrop.adjust_pstomatal_to_eto

AquaCrop.adjust_simperiod!

AquaCrop.adjust_size_compartments!

AquaCrop.adjust_swc_rootzone!

AquaCrop.adjust_theta_initial!

AquaCrop.adjust_year_perennials!

AquaCrop.adjusted_ks_sto_to_ecsw

AquaCrop.adjusted_rooting_depth!

AquaCrop.advance_one_time_step!

AquaCrop.bm_range

AquaCrop.bnormalized

AquaCrop.budget_module!

AquaCrop.calculate_adjusted_fc!

AquaCrop.calculate_capillary_rise!

AquaCrop.calculate_delta_theta

AquaCrop.calculate_drainage!

AquaCrop.calculate_effective_rainfall!

AquaCrop.calculate_etpot

AquaCrop.calculate_evaporation_surface_water!

AquaCrop.calculate_extra_runoff!

AquaCrop.calculate_factor

AquaCrop.calculate_infiltration!

AquaCrop.calculate_irrigation!

AquaCrop.calculate_relative_wetness_topsoil

AquaCrop.calculate_rootfraction_compartment!

AquaCrop.calculate_runoff!

AquaCrop.calculate_salt_content!

AquaCrop.calculate_salt_mobility!

AquaCrop.calculate_sink_values!

AquaCrop.calculate_soil_evaporation_stage1!

AquaCrop.calculate_soil_evaporation_stage2!

AquaCrop.calculate_surfacestorage!

AquaCrop.calculate_theta

AquaCrop.calculate_theta_critical

AquaCrop.calculate_transpiration!

AquaCrop.calculate_weighting_factors!

AquaCrop.canopy_cover_no_stress_days_sf

AquaCrop.canopy_cover_no_stress_gddays_sf

AquaCrop.canopy_cover_no_stress_sf

AquaCrop.cc_at_gdd

AquaCrop.cc_at_gddtime

AquaCrop.cc_at_time

AquaCrop.cc_multiplier_weed

AquaCrop.cc_multiplier_weed_adjusted

AquaCrop.cci_no_water_stress_sf

AquaCrop.ccini_total_from_time_to_ccini

AquaCrop.ccx_salt_stress_relationship!

AquaCrop.check_drain_sum

AquaCrop.check_files_in_project!

AquaCrop.check_for_keep_swc

AquaCrop.check_for_print!

AquaCrop.check_for_watertable_in_profile

AquaCrop.check_germination!

AquaCrop.check_water_salt_balance!

AquaCrop.checkget_gvar_file

AquaCrop.checkget_projectfiles_file

AquaCrop.checkget_resultsparameters_file

AquaCrop.close_climate!

AquaCrop.close_irrigation!

AquaCrop.close_management!

AquaCrop.co2_for_simulation_period

AquaCrop.complete_climate_description!

AquaCrop.complete_crop_description!

AquaCrop.complete_profile_description!

AquaCrop.concentrate_salts!

AquaCrop.correction_anaeroby

AquaCrop.create_daily_climfiles!

AquaCrop.create_eval_data!

AquaCrop.crop_stress_parameters_soil_fertility

AquaCrop.crop_stress_parameters_soil_salinity

AquaCrop.dailyupdate!

AquaCrop.days_to_reach_cc_with_given_cgc

AquaCrop.default_filepaths

AquaCrop.degrees_day

AquaCrop.derive_smax_top_bottom

AquaCrop.designate_soillayer_to_compartments!

AquaCrop.determinate_coeffcapillaryrise!

AquaCrop.determinate_soilclass!

AquaCrop.determine_biomass_and_yield!

AquaCrop.determine_cci!

AquaCrop.determine_cci_gdd!

AquaCrop.determine_ccx_adjusted_cdc

AquaCrop.determine_ccx_adjusted_gdd

AquaCrop.determine_cdc_adjusted_water_stress

AquaCrop.determine_cgc_adjusted

AquaCrop.determine_cni_and_iii

AquaCrop.determine_date

AquaCrop.determine_day_nr

AquaCrop.determine_gddcdc_adjusted_water_stress

AquaCrop.determine_gddcgc_adjusted

AquaCrop.determine_growth_stage!

AquaCrop.determine_length_growth_stages

AquaCrop.determine_linked_simday1!

AquaCrop.determine_nrand_thickness_compartments!

AquaCrop.determine_potential_biomass!

AquaCrop.determine_root_zone_anaero_conditions

AquaCrop.determine_root_zone_salt_content!

AquaCrop.determine_root_zone_wc!

AquaCrop.determine_salt_content!

AquaCrop.ececomp

AquaCrop.ecswcomp

AquaCrop.effect_soil_fertility_salinity_stress!

AquaCrop.extract_water_from_evap_layer!

AquaCrop.fadjusted_for_co2

AquaCrop.feedback_cc!

AquaCrop.file_management!

AquaCrop.finalize_run1!

AquaCrop.finalize_run2!

AquaCrop.find_values

AquaCrop.flush_output_dayout!

AquaCrop.flush_output_etodatasim!

AquaCrop.flush_output_evaldataout!

AquaCrop.flush_output_harvestsout!

AquaCrop.flush_output_logger!

AquaCrop.flush_output_raindatasim!

AquaCrop.flush_output_seasonout!

AquaCrop.flush_output_tcropsim!

AquaCrop.flush_output_tempdatasim!

AquaCrop.fraction_flowering

AquaCrop.fraction_period

AquaCrop.from_gravelmass_to_gravelvol

AquaCrop.full_undefined_record

AquaCrop.gddcdc_to_cdc

AquaCrop.get_cdc_adjusted_no_stress_new

AquaCrop.get_day_switch_to_linear

AquaCrop.get_decade_eto_dataset!

AquaCrop.get_decade_rain_dataset!

AquaCrop.get_decade_temperature_dataset!

AquaCrop.get_gddcdc_adjusted_no_stress

AquaCrop.get_gwt_set!

AquaCrop.get_interpolation_parameters

AquaCrop.get_irri_param!

AquaCrop.get_limits_evap_layer

AquaCrop.get_monthly_eto_dataset!

AquaCrop.get_monthly_rain_dataset!

AquaCrop.get_monthly_temperature_dataset!

AquaCrop.get_new_ccx_and_cdc

AquaCrop.get_new_ccx_and_gddcdc

AquaCrop.get_next_harvest!

AquaCrop.get_parameters

AquaCrop.get_potvalsf

AquaCrop.get_project_type

AquaCrop.get_results_parameters

AquaCrop.get_set_of_three_eto

AquaCrop.get_set_of_three_months_eto

AquaCrop.get_set_of_three_months_rain

AquaCrop.get_set_of_three_months_temperature

AquaCrop.get_set_of_three_temperature

AquaCrop.get_sumgdd_before_simulation!

AquaCrop.get_weed_rc

AquaCrop.get_z_and_ec_gwt!

AquaCrop.getks

AquaCrop.getparameter

AquaCrop.growing_degree_days

AquaCrop.harvest_index_day

AquaCrop.harvest_index_growth_coefficient

AquaCrop.hi_multiplier

AquaCrop.horizontal_inflow_gwtable!

AquaCrop.initialize_climate!

AquaCrop.initialize_project

AquaCrop.initialize_project_filename

AquaCrop.initialize_project_input

AquaCrop.initialize_run_part1!

AquaCrop.initialize_run_part2!

AquaCrop.initialize_settings

AquaCrop.initialize_simulation_run_part1!

AquaCrop.initialize_simulation_run_part2!

AquaCrop.initialize_the_program

AquaCrop.initialize_transfer_assimilates!

AquaCrop.irri_manual!

AquaCrop.irri_out_season

AquaCrop.isleapyear

AquaCrop.ks_any

AquaCrop.ks_salinity

AquaCrop.ks_temperature

AquaCrop.length_canopy_decline

AquaCrop.load_clim!

AquaCrop.load_crop!

AquaCrop.load_groundwater!

AquaCrop.load_gvars_from_toml!

AquaCrop.load_initial_conditions!

AquaCrop.load_irri_schedule_info!

AquaCrop.load_management!

AquaCrop.load_offseason!

AquaCrop.load_profile

AquaCrop.load_profile_processing!

AquaCrop.load_program_parameters_project_plugin!

AquaCrop.load_simulation_project!

AquaCrop.max_crat_depth

AquaCrop.mixing!

AquaCrop.move_salt_to!

AquaCrop.multiplier_cco_self_thinning

AquaCrop.multiplier_ccx_self_thinning

AquaCrop.no_adjustment

AquaCrop.no_effect_stress!

AquaCrop.no_irrigation!

AquaCrop.open_climfiles_and_get_data_firstday!

AquaCrop.open_harvest_info!

AquaCrop.open_irrigation_file!

AquaCrop.prepare_stage1!

AquaCrop.prepare_stage2!

AquaCrop.read_climate_nextday!

AquaCrop.read_output_from_dayout

AquaCrop.read_output_from_etodatasim

AquaCrop.read_output_from_evaldataout

AquaCrop.read_output_from_harvestsout

AquaCrop.read_output_from_logger

AquaCrop.read_output_from_raindatasim

AquaCrop.read_output_from_seasonout

AquaCrop.read_output_from_tcropsim

AquaCrop.read_output_from_tempdatasim

AquaCrop.read_temperature_file!

AquaCrop.record_harvest!

AquaCrop.relationships_for_fertility_and_salt_stress!

AquaCrop.required_gdd

AquaCrop.required_time_new

AquaCrop.reset_previous_sum!

AquaCrop.reset_swc_to_fc!

AquaCrop.root_max_in_soil_profile

AquaCrop.run_simulation!

AquaCrop.salt_solution_deposit!

AquaCrop.salt_transport_factor

AquaCrop.seasonal_sum_of_kcpot

AquaCrop.set_clim_data!

AquaCrop.set_daynr_to_yundef

AquaCrop.set_gdd_variables_nextday!

AquaCrop.setparameter!

AquaCrop.soil_evaporation_reduction_coefficient

AquaCrop.start_outputs

AquaCrop.start_the_program

AquaCrop.stress_biomass_relationship!

AquaCrop.sum_calendar_days

AquaCrop.surface_transpiration!

AquaCrop.swcz_soil

AquaCrop.tau_from_ksat

AquaCrop.temperature_file_covering_crop_period!

AquaCrop.time_root_function

AquaCrop.time_to_cc_ini

AquaCrop.time_to_max_canopy_sf

AquaCrop.translate_inilayers_to_swprofile!

AquaCrop.translate_inipoints_to_swprofile!

AquaCrop.wc_evap_layer

AquaCrop.write_daily_results!

AquaCrop.write_evaluation_data!

AquaCrop.write_intermediate_period!

AquaCrop.write_sim_period!

AquaCrop.write_the_results!

AquaCrop.write_title_part1_mult_results!

AquaCrop.year_weighing_factor

AquaCrop.zr_adjusted_to_restrictive_layers

AquaCrop.ton — Constant

ton

metric tonelad 1 ton = 1000 kg Dimension: 𝐌

AquaCrop.AbstractParametersContainer — Type

AbstractParametersContainer

AquaCrop.CompartmentIndividual — Type

compartment = CompartmentIndividual()

AquaCrop.FieldAquaCrop — Type

FieldAquaCrop

Has all the data for the simulation of AquaCrop

AquaCrop.FieldAquaCrop — Method

FieldAquaCrop(inputDataDir::String)

Starts the struct FieldAquaCrop that has all the data for the simulation of AquaCrop

AquaCrop.ParametersContainer — Type

pc = ParametersContainer(T)

Contains parameters in a Dictionary of type Symbol=>T. it is used for parameters without a type in the original AquaCrop.f90

AquaCrop.ProjectInputType — Type

projectinput = ProjectInputType()

Container for project file input data.

AquaCrop.RepAssimilates — Type

assimilates = RepAssimilates()

set from defaultcropsoil.f90:274

AquaCrop.RepClim — Type

record_clim = RepClim()

AquaCrop.RepContent — Type

content = RepContent()

AquaCrop.RepCrop — Type

crop = RepCrop()

set from defaultcropsoil.f90:140

AquaCrop.RepCropFileSet — Type

crop_file_set = RepCropFileSet()

AquaCrop.RepCutInfoRecord — Type

cut_info = RepCutInfoRecord()

AquaCrop.RepCuttings — Type

cuttings = RepCuttings()

set from global.f90:3338

AquaCrop.RepDayEventDbl — Type

a = RepDayEventDbl()

AquaCrop.RepDayEventInt — Type

a = RepDayEventInt()

set from global.f90:2838

AquaCrop.RepEffectStress — Type

effectstress = RepEffectStress()

AquaCrop.RepEffectiveRain — Type

effectiverain = RepEffectiveRain()

set from initialsettings.f90:244

AquaCrop.RepEndSeason — Type

endseason = RepEndSeason()

AquaCrop.RepFileOK — Type

fileok = RepFileOK()

AquaCrop.RepGwTable — Type

gwtable = RepGwTable()

AquaCrop.RepIniSWC — Type

iniswc = RepIniSWC()

AquaCrop.RepIrriECw — Type

irriecw = RepIrriECw()

set from global.f90:2838

AquaCrop.RepIrriInfoRecord — Type

irri_info = RepIrriInfoRecord()

AquaCrop.RepManag — Type

management = RepManag()

set from global.f90:3311

AquaCrop.RepOnset — Type

onset = RepOnset()

set from initialsettings.f90:464

AquaCrop.RepParam — Type

simulparams = RepParams()

contains the simulation parameters.

set from initialsettings.f90:216

AquaCrop.RepPerennialPeriod — Type

perennial_period = RepPerennialPeriod()

AquaCrop.RepPlotPar — Type

plotvarcorp = RepPlotPar()

we set these values since we need them for rootunit.f90:37 AdjustedRootingDepth

AquaCrop.RepRootZoneSalt — Type

root_zone_salt = RepRootZoneSalt()

AquaCrop.RepRootZoneWC — Type

root_zone_wc = RepRootZoneWC()

AquaCrop.RepShapes — Type

shape = RepShapes()

set from defaultcropsoil.f90:175

AquaCrop.RepSim — Type

simulation = RepSim()

set from global.f90:7694

AquaCrop.RepSoil — Type

soil = RepSoil()

set from defaultcropsoil.f90:289

AquaCrop.RepStorage — Type

storage = RepStorage()

AquaCrop.RepStressTot — Type

stresstot = RepStressTot()

AquaCrop.RepSum — Type

summ = RepSum()

set from global.f90:7153

AquaCrop.RepTransfer — Type

transfer = RepTransfer()

AquaCrop.SoilLayerIndividual — Type

soillayer = SoilLayerIndividual()

creates a soil layer with a given soil class.

set from defaultcropsoil.f90:289

AquaCrop.StressIndexesBio — Type

stress_indexes = StressIndexesBio()

AquaCrop.StressIndexesSalt — Type

stress_indexes = StressIndexesSalt()

AquaCrop.active_cells — Method

celi = active_cells(compartment::CompartmentIndividual, soil_layers::Vector{SoilLayerIndividual})

globa.f90:4241

AquaCrop.actual_rooting_depth — Method

zr = actual_rooting_depth(dap, l0, lzmax, l1234, gddl0, gddlzmax, 
                          sumgdd, zmin, zmax, shapefactor, typedays, gvars)

global.f90:7262 must set simulation.SCor = 1 before calling this function

AquaCrop.actual_rooting_depth_days — Method

ardd = actual_rooting_depth_days(dap, l0, lzmax, l1234, zmin, zmax, shapefactor, gvars)

global.f90:7306

AquaCrop.actual_rooting_depth_gddays — Method

ardg = actual_rooting_depth_gddays(dap, l1234, gddl0, gddlzmax, sumgdd, zmin, zmax, shapefactor, gvars)

global.f90:7346

AquaCrop.actualize_with_dict! — Method

actualize_with_dict!(obj::T, aux::AbstractDict) where T<:AbstractParametersContainer

AquaCrop.add_output_in_dayout! — Method

add_output_in_dayout(outputs, arr)

AquaCrop.add_output_in_etodatasim! — Method

add_output_in_etodatasim!(outputs, eto::T) where T<:Number

AquaCrop.add_output_in_evaldataout! — Method

add_output_in_evaldataout!(outputs, arr)

AquaCrop.add_output_in_harvestsout! — Method

add_output_in_harvestsout!(outputs, arr)

AquaCrop.add_output_in_logger! — Method

add_output_in_logger!(outputs, aux::String)

AquaCrop.add_output_in_raindatasim! — Method

add_output_in_raindatasim!(outputs, rain::T) where T<:Number

AquaCrop.add_output_in_seasonout! — Method

add_output_in_seasonout!(outputs, arr)

AquaCrop.add_output_in_tcropsim! — Method

add_output_in_tcropsim!(outputs, tlow::T, thigh::T) where T<:Number

AquaCrop.add_output_in_tempdatasim! — Method

add_output_in_tempdatasim!(outputs, tlow::T, thigh::T) where T<:Number

AquaCrop.adjust_calendar_crop! — Method

adjust_calendar_crop!(gvars)

tempprocessing.f90:1467

AquaCrop.adjust_calendar_days! — Method

adjust_calendar_days!(gvars, iscgcgiven)

tempprocessing.f90:1327

AquaCrop.adjust_climrecord_to! — Method

adjust_climrecord_to!(clim_record::RepClim, cdayn)

global.f90:6163

AquaCrop.adjust_compartments! — Method

adjust_compartments!(gvars)

run.f90:6619

AquaCrop.adjust_crop_file_parameters! — Method

adjust_crop_file_parameters!(gvars)

tempprocessing.f90:1882

AquaCrop.adjust_crop_year_to_climfile! — Method

adjust_crop_year_to_climfile!(crop::RepCrop, clim_file, clim_record)

global.f90:6447

AquaCrop.adjust_decade_month_and_year — Method

decfile, mfile, yfile = adjust_decade_month_and_year(decfile, mfile, yfile)

tempprocessing.f90:293

AquaCrop.adjust_epot_mulch_wetted_surface! — Method

adjust_epot_mulch_wetted_surface!(gvars)

simul.f90:4260

AquaCrop.adjust_for_watertable! — Method

adjust_for_watertable!(gvars)

run.f90:3423

AquaCrop.adjust_month — Method

cimin, cimax = adjust_month(tlow, thigh)

this comes from read_month tempprocessing.f90:837

AquaCrop.adjust_month_and_year — Method

mfile, yfile = adjust_month_and_year(mfile, yfile)

tempprocessing.f90:284

AquaCrop.adjust_onset_search_period! — Method

adjust_onset_search_period!(gvars)

global.f90:4214

AquaCrop.adjust_pleaf_to_eto — Method

pleafulact, pleafllact = adjust_pleaf_to_eto(etomean, crop, simulparam)

simul.f90:424

AquaCrop.adjust_psenescence_to_eto — Method

psenact = adjust_psenescence_to_eto(etomean, timesenescence, withbeta, crop, simulparam)

simul.f90:1106

AquaCrop.adjust_pstomatal_to_eto — Method

pstomatulact = adjust_pstomatal_to_eto(meaneto, crop, simulparam)

simul.f90:1084

AquaCrop.adjust_simperiod! — Method

adjust_simperiod!(gvars, projectinput::ProjectInputType)

global.f90:4692

AquaCrop.adjust_size_compartments! — Method

adjust_size_compartments!(gvars, cropzx)

global.f90:6563

AquaCrop.adjust_swc_rootzone! — Method

adjust_swc_rootzone!(gvars, lvars)

run.f90:6348

AquaCrop.adjust_theta_initial! — Method

adjust_theta_initial!(gvars, prevnrcomp, prevthickcomp, prevvolprcomp, prevecdscomp)

global.f90:5852

AquaCrop.adjust_year_perennials! — Method

adjust_year_perennials!(crop::RepCrop, theyearseason)

global.f90:4690

AquaCrop.adjusted_ks_sto_to_ecsw — Method

ksstoout = adjusted_ks_sto_to_ecsw(ecemin, ecemax, responseecsw, ecei, 
        ecswi, ecswfci, wrel, coeffb0salt, coeffb1salt, coeffb2salt, ksstoin, simulation)

global.f90:2321

AquaCrop.adjusted_rooting_depth! — Method

adjusted_rooting_depth!(gvars)

rootunit.f90:37

AquaCrop.advance_one_time_step! — Method

advance_one_time_step!(outputs, gvars, lvars, projectinput::ProjectInputType, nrrun)

run.f90:6747

AquaCrop.bm_range — Method

bmr = bm_range(hiadj)

global.f90:2278

AquaCrop.bnormalized — Method

sumbnor = bnormalized(outputs, thedaystoccini, thegddaystoccini,
        l0, l12, l12sf, l123, l1234, lflor, 
        gddl0, gddl12, gddl12sf, gddl123, gddl1234, 
        wpyield, daysyieldformation, tswitch, cco, ccx, 
        cgc, gddcgc, cdc, gddcdc, kctop, kcdeclageing, 
        cceffectprocent, wpbio, theco2, tbase, tupper, 
        tdaymin, tdaymax, gdtransplow, ratdgdd, sumkctop, 
        stressinpercent, strresredcgc, strresredccx, strresredwp, 
        strresredkssto, weedstress, deltaweedstress, strrescdecline, 
        shapefweed, themodecycle, fertilitystresson,
        simulation, simulparam, management, crop)

tempprocessing.f90:2586 note that we must do simulation.DelayedDays = 0 before calling this function

AquaCrop.budget_module! — Method

budget_module!(gvars, lvars, virtualtimecc, sumgddadjcc)

simul.f90:5417

AquaCrop.calculate_adjusted_fc! — Method

calculate_adjusted_fc!(compartadj::Vector{CompartmentIndividual}, soil_layers::Vector{SoilLayerIndividual}, depthaquifer)

global.f90:4141

AquaCrop.calculate_capillary_rise! — Method

calculate_capillary_rise!(gvars)

simul.f90:2060

AquaCrop.calculate_delta_theta — Method

deltax = calculate_delta_theta(theta_in, thetaadjfc, nrlayer, soil_layers)

simul.f90:1570

AquaCrop.calculate_drainage! — Method

calculate_drainage!(gvars)

simul.f90:1627

AquaCrop.calculate_effective_rainfall! — Method

calculate_effective_rainfall!(gvars, lvars)

simul.f90:1986

AquaCrop.calculate_etpot — Method

tpotval, epotval = calculate_etpot(dap, l0, l12, l123, lharvest, daylastcut, cci, etoval, kcval, kcdeclineval, ccx, ccxwithered, cceffectprocent, co2i, gddayi, tempgdtransplow, simulation, simulparam)

global.f90:7888

AquaCrop.calculate_evaporation_surface_water! — Method

calculate_evaporation_surface_water!(gvars)

simul.f90:4236

AquaCrop.calculate_extra_runoff! — Method

calculate_extra_runoff!(gvars, lvars)

simul.f90:2720

AquaCrop.calculate_factor — Method

cf = calculate_factor(layeri, compi, soil_layers, compartments)

simu.f90:3212

AquaCrop.calculate_infiltration! — Method

calculate_infiltration!(gvars, lvars)

simul.f90:2820

AquaCrop.calculate_irrigation! — Method

calculate_irrigation!(gvars, lvars, targettimeval, targetdepthval)

simul.f90:1940

AquaCrop.calculate_relative_wetness_topsoil — Method

sumi = calculate_relative_wetness_topsoil(maxdepth, compartments, soil_layers)

simul.f90:1901

AquaCrop.calculate_rootfraction_compartment! — Method

calculate_rootfraction_compartment!(compartments, rootingdepth)

simul.f90:1375

AquaCrop.calculate_runoff! — Method

calculate_runoff!(gvars)

simul.f90:1852

AquaCrop.calculate_salt_content! — Method

calculate_salt_content!(gvars, lvars, dayi)

simul.f90:2329

AquaCrop.calculate_salt_mobility! — Method

calculate_salt_mobility!(soillayer::SoilLayerIndividual, saltdiffusion)

sets the salt mobility

global.f90:7500

AquaCrop.calculate_sink_values! — Method

calculate_sink_values!(compartments, rootingdepth, irrimode, crop, simulation)

simul.f90:1407

AquaCrop.calculate_soil_evaporation_stage1! — Method

calculate_soil_evaporation_stage1!(gvars)

simul.f90:4439

AquaCrop.calculate_soil_evaporation_stage2! — Method

calculate_soil_evaporation_stage2!(gvars)

simul.f90:4458

AquaCrop.calculate_surfacestorage! — Method

calculate_surfacestorage!(gvars, lvars, dayi)

simul.f90:2756

AquaCrop.calculate_theta — Method

thetax = calculate_theta(delta_theta, thetaadjfc, nrlayer, soil_layers)

simul.f90:1598

AquaCrop.calculate_theta_critical — Method

theta = calculate_theta_critical(layeri, soil_layers, crop)

simul.f90:1362

AquaCrop.calculate_transpiration! — Method

calculate_transpiration!(gvars, tpot, coeffb0salt, coeffb1salt, coeffb2salt)

simul.f90:1156

AquaCrop.calculate_weighting_factors! — Method

calculate_weighting_factors!(depth, compartments)

simul.f90:1819

AquaCrop.canopy_cover_no_stress_days_sf — Method

cc = canopy_cover_no_stress_days_sf(dap, l0, l123,
   lmaturity, cco, ccx, cgc, cdc, sfredcgc, sfredccx, simulation)

global.f90:1333

AquaCrop.canopy_cover_no_stress_gddays_sf — Method

cc = canopy_cover_no_stress_gddays_sf(gddl0, gddl123, gddlmaturity, sumgdd, 
    cco, ccx, gddcgc, gddcdc, sfredcgc, sfredccx)

global.f90:2670

AquaCrop.canopy_cover_no_stress_sf — Method

canopycovernostresssf = canopy_cover_no_stress_sf(dap, l0, l123, 
   lmaturity, gddl0, gddl123, gddlmaturity, cco, ccx,
   cgc, cdc, gddcgc, gddcdc, sumgdd, typedays, sfredcgc, sfredccx, simulation)

global.f90:1299

AquaCrop.cc_at_gdd — Method

cci = cc_at_gdd(gddi, ccoin, gddcgcin, ccxin)

global.f90:2654

AquaCrop.cc_at_gddtime — Method

cci = cc_at_gddtime(gddtfictive, ccogiven, gddcgcgiven, ccxgiven)

simul.f90:3863

AquaCrop.cc_at_time — Method

cci = cc_at_time(dayi, ccoin, cgcin, ccxin)

global.f90:2371

AquaCrop.cc_multiplier_weed — Method

fweed = cc_multiplier_weed(procentweedcover, ccxcrop, fshapeweed)

global.f90:2180

AquaCrop.cc_multiplier_weed_adjusted — Method

fweedi, rcadj = cc_multiplier_weed_adjusted(procentweedcover, ccxcrop, fshapeweed, fccx, yeari, mweedadj, cropsubkind)

global.f90:2204

AquaCrop.cci_no_water_stress_sf — Method

cci = cci_no_water_stress_sf(dayi, l0, l12sf, l123, l1234, gddl0,
                gddl12sf, gddl123, gddl1234, cco, ccx, cgc, gddcgc, cdc, gddcdc, sumgdd,
                ratdgdd, sfredcgc, sfredccx, sfcdecline, themodecycle, simulation)

global.f90:1391

AquaCrop.ccini_total_from_time_to_ccini — Method

ccini = ccini_total_from_time_to_ccini(tempdaystoccini, tempgddaystoccini, 
                                        l0, l12, l12sf, l123, l1234, gddl0, 
                                        gddl12, gddl12sf, gddl123, 
                                        gddl1234, cco, ccx, cgc, gddcgc, 
                                        cdc, gddcdc, ratdgdd, sfredcgc, 
                                        sfredccx, sfcdecline, fweed, 
                                        themodecycle, simulation)

global.f90:6371

AquaCrop.ccx_salt_stress_relationship! — Method

ccx_salt_stress_relationship!(outputs, gvars)

tempprocessing.f90:3277

AquaCrop.check_drain_sum — Method

drainsum, excess = check_drain_sum(layeri, drainsum, excess, soil_layers)

simul.f90:1806

AquaCrop.check_files_in_project! — Method

check_files_in_project!(fileok::RepFileOK, allok, input::ProjectInputType)

global.f90:7195

AquaCrop.check_for_keep_swc — Method

runwithkeepswc, constzrxforrun = checkforkeep_swc(outputs, projectinput::Vector{ProjectInputType}, filepaths, gvars; kwargs...)

global.f90:6643

AquaCrop.check_for_print! — Method

check_for_print!(outputs, gvars)

run.f90:3476

AquaCrop.check_for_watertable_in_profile — Method

logi = check_for_watertable_in_profile(profilecomp::Vector{CompartmentIndividual}, depthgwtmeter)

global.f90:1540

AquaCrop.check_germination! — Method

check_germination!(gvars)

simul.f90:1130

AquaCrop.check_water_salt_balance! — Method

check_water_salt_balance!(gvars, lvars, dayi, control)

simul.f90:2191

AquaCrop.checkget_gvar_file — Method

checkget_gvar_file(outputs, filename)

we check if the file exists, if not we give the default file path

AquaCrop.checkget_projectfiles_file — Method

checkget_projectfiles_file(outputs, filename)

we check if the file exists, if not we give the default file path

AquaCrop.checkget_resultsparameters_file — Method

checkget_resultsparameters_file(outputs, filename)

we check if the file exists, if not we give the default file path

AquaCrop.close_climate! — Method

close_climate!(outputs, gvars; kwargs...)

delete the climate data from gvars arrays and outputs arrays

AquaCrop.close_irrigation! — Method

close_irrigation!(gvars; kwargs...)

delete the irrigation data from gvars arrays

AquaCrop.close_management! — Method

close_management!(gvars; kwargs...)

delete the management data from gvars arrays

AquaCrop.co2_for_simulation_period — Method

co2forsimulationperiod = co2_for_simulation_period(co2_file, fromdaynr, todaynr)

global.f90:3114

AquaCrop.complete_climate_description! — Method

complete_climate_description!(record::RepClim)

global.f90:8223

AquaCrop.complete_crop_description! — Method

complete_crop_description!(crop::RepCrop, simulation::RepSim, management::RepManag)

global.f90:5624

AquaCrop.complete_profile_description! — Method

complete_profile_description!(soil_layers::Vector{SoilLayerIndividual}, 
        compartments::Vector{CompartmentIndividual}, simulation::RepSim, total_water_content::RepContent)

global.f90:7561

AquaCrop.concentrate_salts! — Method

concentrate_salts!(gvars)

simul.f90:4343

AquaCrop.correction_anaeroby — Method

alfa, dayanaero = correction_anaeroby(compartment, alfa, daysubmerged, simulparam, soil_layers, crop)

simul.f90:1453 note that we must do compartment.DayAnaero = dayanaero after calling this function

AquaCrop.create_daily_climfiles! — Method

create_daily_climfiles(outputs, gvars; kwargs...)

run.f90:5592

AquaCrop.create_eval_data! — Method

create_eval_data!(gvars)

run.f90:5451

AquaCrop.crop_stress_parameters_soil_fertility — Method

stressout = crop_stress_parameters_soil_fertility(cropsresp::RepShapes, stresslevel)

global.f90:1231

AquaCrop.crop_stress_parameters_soil_salinity — Method

stress_response = crop_stress_parameters_soil_salinity(ccxred, ccdistortion, 
         cco, ccx, cgc, gddcgc, cropdeterm, l12, lflor, 
         lengthflor, l123, gddl12, gddlflor, gddlengthflor, 
         gddl123, themodecycle)

tempprocessing.f90:1643

AquaCrop.dailyupdate! — Method

dailyupdate!(fieldcrop::AbstractFieldCrop)

Updates the FieldCrop by one day

AquaCrop.days_to_reach_cc_with_given_cgc — Method

daystoresult = days_to_reach_cc_with_given_cgc(cctoreach, ccoval, ccxval, cgcval, l0)

global.f90:1809

AquaCrop.default_filepaths — Method

fl = default_filepaths(parentdir::AbstractString; kwargs...)

sets the default directories for the input files.

startunit.f90:420

AquaCrop.degrees_day — Method

dgrd = degrees_day(tbase, tupper, tdaymin, tdaymax, gddselectedmethod)

global.f90:2419

AquaCrop.derive_smax_top_bottom — Method

sxtop, sxbot = derive_smax_top_bottom(crop::RepCrop)

global.f90:1944

AquaCrop.designate_soillayer_to_compartments! — Method

designate_soillayer_to_compartments!(compartments::Vector{CompartmentIndividual}, soil_layers::Vector{SoilLayerIndividual})

global.f90:7393

AquaCrop.determinate_coeffcapillaryrise! — Method

determinate_coeffcapillaryrise!(soillayer::SoilLayerIndividual)

sets the coefficients for capillary rise of soillayer considering its own data.

global.f90:4034

AquaCrop.determinate_soilclass! — Method

determinate_soilclass!(soillayer::SoilLayerIndividual)

sets the soil class of soillayer considering its own data.

global.f90:1909

AquaCrop.determine_biomass_and_yield! — Method

determine_biomass_and_yield!(gvars, lvars, sumgddadjcc, virtualtimecc)

simul.f90:528

AquaCrop.determine_cci! — Method

determine_cci!(gvars, ccxtotal, ccototal, fracassim, 
                        mobilizationon, storageon, tadj, virtualtimecc, 
                        cdctotal, dayfraction, 
                        gddcdctotal)

simul.f90:4611

AquaCrop.determine_cci_gdd! — Method

determine_cci_gdd!(gvars, ccxtotal, ccototal, 
                        fracassim, mobilizationon, 
                        storageon, sumgddadjcc, virtualtimecc, 
                        cdctotal, gddayfraction, 
                        gddayi, gddcdctotal, gddtadj)

simul.f90:3234

AquaCrop.determine_ccx_adjusted_cdc — Method

ccxadjusted = determine_ccx_adjusted_cdc(cciprev, virtualtimecc, cgcadjusted, ccxsf, tfinalccx, crop)

simul.f90:5310

AquaCrop.determine_ccx_adjusted_gdd — Method

ccxadjusted = determine_ccx_adjusted_gdd(cciprev, ccxsf, gddcgcadjusted, gddayi, sumgddadjcc, gddtfinalccx, cco)

simul.f90:3882

AquaCrop.determine_cdc_adjusted_water_stress — Method

cdcadjusted, kssen, stresssenescence = determine_cdc_adjusted_water_stress(eto, timesenescence, cdctotal, ccxsfcd, ccxtotal,
                                                                           root_zone_wc, crop, simulparam, simulation)

simul.f90:5222

AquaCrop.determine_cgc_adjusted — Method

cgcadjusted, stressleaf = determine_cgc_adjusted(cgcsf, pleafllact, 
                                                 root_zone_wc, crop, simulation)

simul.f90:5171

AquaCrop.determine_cni_and_iii — Method

cn1, cn3 = determine_cni_and_iii(cn2)

global.f90:2481

AquaCrop.determine_date — Method

dayi, monthi, yeari = determine_date(dar_nr)

global.f90:2397

AquaCrop.determine_day_nr — Method

day_nr = determine_day_nr(dayi, monthi, yeari)

global.f90:2387

AquaCrop.determine_gddcdc_adjusted_water_stress — Method

gddcdcadjusted, kssen, stresssenescence = determine_gddcdc_adjusted_water_stress(eto, timesenescence, gddcdctotal, ccxsfcd, ccxtotal,
                                                        root_zone_wc, crop, simulparam, simulation)

simul.f90:3917

AquaCrop.determine_gddcgc_adjusted — Method

gddcgcadjusted, stressleaf = determine_gddcgc_adjusted(cgcgddsf, pleafllact, 
                                                       root_zone_wc, crop, simulation)

simul.f90:3788

AquaCrop.determine_growth_stage! — Method

determine_growth_stage!(gvars, dayi, cciprev)

run.f90:4075

AquaCrop.determine_length_growth_stages — Method

length123, stlength, length12, cgcval = determine_length_growth_stages(ccoval, ccxval, cdcval, l0, totallength, 
                                                                    cgcgiven, thedaystoccini, theplanting, 
                                                                    length123, stlength, length12, cgcval)

global.f90:1644

AquaCrop.determine_linked_simday1! — Method

determine_linked_simday1!(simulation::RepSim, crop::RepCrop, clim_record::RepClim, clim_file)

global.f90:4677

AquaCrop.determine_nrand_thickness_compartments! — Method

determine_nrand_thickness_compartments!(compartments::Vector{CompartmentIndividual}, soil_layers::Vector{SoilLayerIndividual}, compdefthick)

global.f90:4063

AquaCrop.determine_potential_biomass! — Method

determine_potential_biomass!(gvars, virtualtimecc, sumgddadjcc)

simul.f90:453

AquaCrop.determine_root_zone_anaero_conditions — Method

redfact, dayanaero = determine_root_zone_anaero_conditions(wsat, wact, anaevol, zr, simulation, simulparam)

simul.f90:1490 note that we must do simulation.DayAnaero = dayanaero after calling this function

AquaCrop.determine_root_zone_salt_content! — Method

determine_root_zone_salt_content!(gvars, rootingdepth)

global.f90:4089

AquaCrop.determine_root_zone_wc! — Method

determine_root_zone_wc!(gvars, rootingdepth)

global.f90:7750

AquaCrop.determine_salt_content! — Method

determine_salt_content!(compartment::CompartmentIndividual, soil_layers::Vector{SoilLayerIndividual}, simulparam::RepParam)

global.f90:4258

AquaCrop.ececomp — Method

ece = ececomp(compartment::CompartmentIndividual, gvars)

global.f90:2523

AquaCrop.ecswcomp — Method

ecs = ecswcomp(compartment::CompartmentIndividual, atfc, gvars)

global.f90:2547

AquaCrop.effect_soil_fertility_salinity_stress! — Method

effect_soil_fertility_salinity_stress!(gvars, stresssfadjnew, coeffb0salt, 
                                             coeffb1salt, coeffb2salt, 
                                             nrdaygrow, stresstotsaltprev, 
                                             virtualtimecc)

simul.f90:3985

AquaCrop.extract_water_from_evap_layer! — Method

extract_water_from_evap_layer!(gvars, evaptolose, zact, stg1)

simul.f90:4377

AquaCrop.fadjusted_for_co2 — Method

fadjustedforco2 = fadjusted_for_co2(co21, wpi, percenta)

global.f90:2749

AquaCrop.feedback_cc! — Method

feedback_cc!(gvars)

simul.f90:5348

AquaCrop.file_management! — Method

file_management!(outputs, gvars, projectinput::ProjectInputType, nrrun; kwargs...)

run.f90:7760

AquaCrop.finalize_run1! — Method

finalize_run1!(outputs, gvars, nrrun; kwargs...)

run.f90:7390

AquaCrop.finalize_run2! — Method

finalize_run2!(outputs, gvars; kwargs...)

run.f90:4355

AquaCrop.find_values — Method

zcn, ecdsm = find_values(atdaynr, daynr1, daynr2, z1, ec1, z2, ec2)

global.f90:6140

AquaCrop.flush_output_dayout! — Method

flush_output_dayout!(outputs)

AquaCrop.flush_output_etodatasim! — Method

flush_output_etodatasim!(outputs)

AquaCrop.flush_output_evaldataout! — Method

flush_output_evaldataout!(outputs)

AquaCrop.flush_output_harvestsout! — Method

flush_output_harvestsout!(outputs)

AquaCrop.flush_output_logger! — Method

flush_output_logger!(outputs)

AquaCrop.flush_output_raindatasim! — Method

flush_output_raindatasim!(outputs)

AquaCrop.flush_output_seasonout! — Method

flush_output_seasonout!(outputs)

AquaCrop.flush_output_tcropsim! — Method

flush_output_tcropsim!(outputs)

AquaCrop.flush_output_tempdatasim! — Method

flush_output_tempdatasim!(outputs)

AquaCrop.fraction_flowering — Method

f = fraction_flowering(dayi, crop, simulation)

simul.f90:1025

AquaCrop.fraction_period — Method

fi = fraction_period(diflor, crop)

simul.f90:1050

AquaCrop.from_gravelmass_to_gravelvol — Method

gravelvol = from_gravelmass_to_gravelvol(porositypercent, gravelmasspercent)

calculates the gravel volume of soil layer.

global.f90:1521

AquaCrop.full_undefined_record — Method

logi = full_undefined_record(record::RepClim)

global.f90:2826

AquaCrop.gddcdc_to_cdc — Method

cdc = gddcdc_to_cdc(plantdaynr, d123, gddl123, gddharvest, ccx, gddcdc, tbase, tupper, notempfiletmin, notempfiletmax, gvars)

tempprocessing.f90:1545

AquaCrop.get_cdc_adjusted_no_stress_new — Method

cdcadjusted = get_cdc_adjusted_no_stress_new(ccx, cdc, ccxadjusted)

simul.f90:412

AquaCrop.get_day_switch_to_linear — Method

tswitch, higclinear = get_day_switch_to_linear(himax, dhidt, higc)

global.f90:2988

AquaCrop.get_decade_eto_dataset! — Method

get_decade_eto_dataset!(eto_dataset, daynri, eto_array, eto_record::RepClim)

climprocessing.f90:325

AquaCrop.get_decade_rain_dataset! — Method

get_decade_rain_dataset!(rain_dataset, daynri, rain_array, rain_record::RepClim)

climprocessing.f90:515

AquaCrop.get_decade_temperature_dataset! — Method

get_decade_temperature_dataset!(tmin_dataset, tmax_dataset, daynri, temperature_array, temperature_record::RepClim)

tempprocessing.f90:362

AquaCrop.get_gddcdc_adjusted_no_stress — Method

gddcdcadjusted = get_gddcdc_adjusted_no_stress(ccx, gddcdc, ccxadjusted)

simul.f90:3905

AquaCrop.get_gwt_set! — Method

get_gwt_set!(gvars, parentdir, daynrin)

run.f90:3526

AquaCrop.get_interpolation_parameters — Method

 aover3, bover2, c = get_interpolation_parameters(c1, c2, c3)

tempprocessing.f90:854

AquaCrop.get_irri_param! — Method

get_irri_param!(gvars, lvars)

run.f90:6262

AquaCrop.get_limits_evap_layer — Method

wupper, wlower = get_limits_evap_layer(xproc, simulation, compartments, soil_layers, soil)

simul.f90:4575

AquaCrop.get_monthly_eto_dataset! — Method

get_monthly_eto_dataset!(eto_dataset, daynri, eto_array, eto_record::RepClim)

climprocessing.f90:87

AquaCrop.get_monthly_rain_dataset! — Method

get_monthly_rain_dataset!(rain_dataset, daynri, rain_array, rain_record::RepClim)

climprocessing.f90:604

AquaCrop.get_monthly_temperature_dataset! — Method

get_monthly_temperature_dataset!(tmin_dataset, tmax_dataset, daynri, temperature_array, temperature_record::RepClim)

tempprocessing.f90:596

AquaCrop.get_new_ccx_and_cdc — Method

ccxadjusted, cdcadjusted = get_new_ccx_and_cdc(cciprev, cdc, ccx, virtualtimecc, crop)

simul.f90:5331

AquaCrop.get_new_ccx_and_gddcdc — Method

ccx, gddcdc = get_new_ccx_and_gddcdc(cciprev, gddcdc, ccx, 
                                        sumgddadjcc, gddayi,
                                        crop)

simul.f90:3967

AquaCrop.get_next_harvest! — Method

get_next_harvest!(gvars)

run.f90:3670

AquaCrop.get_parameters — Method

ul, ll, mid = get_parameters(c1, c2, c3)

tempprocessing.f90:580

AquaCrop.get_potvalsf — Method

potvalsf = get_potvalsf(dap, sumgddadjcc, gvars)

run.f90:6468

AquaCrop.get_project_type — Method

pt = get_project_type(theprojectfile; kwargs...)

gets the project type for a given file.

startunit.f90:322

AquaCrop.get_results_parameters — Method

resultsparameters = get_results_parameters(outputs, path::String; kwargs...)

gets all the results parameters in filepaths[:simul].

startunit.f90:426

AquaCrop.get_set_of_three_eto — Method

c1, c2, c3 = get_set_of_three_eto(dayn, deci, monthi, yeari, eto_array, eto_record::RepClim)

climprocessing.f90:393

AquaCrop.get_set_of_three_months_eto — Method

c1, c2, c3, x1, x2, x3, t1 = get_set_of_three_months_eto(monthi, yeari, eto_array, eto_record)

climprocessing.f90:128

AquaCrop.get_set_of_three_months_rain — Method

c1, c2, c3 = get_set_of_three_months_rain(monthi, yeari, rain_array, rain_record)

climprocessing.f90:665

AquaCrop.get_set_of_three_months_temperature — Method

c1min, c2min, c3min, c1max, c2max, c3max, x1, x2, x3, t1 = get_set_of_three_months_temperature(monthi, yeari, temperature_array, temperature_record::RepClim)

tempprocessing.f90:645

AquaCrop.get_set_of_three_temperature — Method

c1min, c1max, c2min, c2max, c3min, c3max = get_set_of_three_temperature(dayn, deci, monthi, yeari, temperature_array, temperature_record::RepClim)

tempprocessing.f90:439

AquaCrop.get_sumgdd_before_simulation! — Method

get_sumgdd_before_simulation!(gvars)

run.f90:3798

AquaCrop.get_weed_rc — Method

weedrcdaycalc, tempweeddeltarc = get_weed_rc(theday, gddayi, fccx, tempweedrcinput, tempweedadj,
                        tempweeddeltarc, l12sf, templ123, gddl12sf, 
                        tempgddl123, themodecycle)

global.f90:1567

AquaCrop.get_z_and_ec_gwt! — Method

get_z_and_ec_gwt!(gvars)

run.f90:6159

AquaCrop.getks — Method

ksi = getks(t0, t1, tin)

global.f90:2021

AquaCrop.getparameter — Method

getparameter(parameterscontainer::ParametersContainer{T}, parameterkey::Symbol)

gets the parameterkey from parameterscontainer

AquaCrop.growing_degree_days — Method

gdd1234 = growing_degree_days(valperiod, firstdayperiod, tbase, tupper, gvars, tdaymin, tdaymax)

tempprocessing.f90:871

AquaCrop.harvest_index_day — Method

hiday, hifinal, percentlagphase = harvest_index_day(dap, daystoflower, himax, dhidt, cci, 
                              ccxadjusted, theccxwithered, 
                              percccxhifinal, tempplanting, 
                              hifinal,
                              crop, simulation)

global.f90:5530

AquaCrop.harvest_index_growth_coefficient — Method

higc = harvest_index_growth_coefficient(himax, dhidt)

global.f90:1860

AquaCrop.hi_multiplier — Method

himultiplier = hi_multiplier(ratiobm, rangebm, hiadj)

global.f90:2295

AquaCrop.horizontal_inflow_gwtable! — Method

horizontal_inflow_gwtable!(gvars, lvars, depthgwtmeter)

simul.f90:5360

AquaCrop.initialize_climate! — Method

initialize_climate!(outputs, gvars; kwargs...)

run.f90:5006

AquaCrop.initialize_project — Method

gvars, projectinput, fileok = initialize_project(outputs, theprojectfile, theprojecttype, filepaths; kwargs...)

startunit.f90:535

AquaCrop.initialize_project_filename — Method

project_filenames = initialize_project_filename(outputs, filepaths; kwargs...)

Gets all the names of the projects files.

startunit.f90:441

AquaCrop.initialize_project_input — Method

projectinput = initialize_project_input(filename, parentdir; kwargs...)

project_input.f90:152

AquaCrop.initialize_run_part1! — Method

initialize_run_part1!(outputs, gvars, projectinput::ProjectInputType; kwargs...)

run.f90:6590

AquaCrop.initialize_run_part2! — Method

initialize_run_part2!(outputs, gvars, projectinput::ProjectInputType, nrun; kwargs...)

run.f90:6672

AquaCrop.initialize_settings — Method

gvars = initialize_settings(outputs, filepaths; kwargs...)

gets the initial settings.

initialsettings.f90:201

AquaCrop.initialize_simulation_run_part1! — Method

initialize_simulation_run_part1!(outputs, gvars, projectinput)

run.f90:4754

AquaCrop.initialize_simulation_run_part2! — Method

initialize_simulation_run_part2!(outputs, gvars, projectinput::ProjectInputType; kwargs...)

run.f90:5018

AquaCrop.initialize_the_program — Method

filepaths, resultsparameters = initialize_the_program(outputs, parentdir; kwargs...)

Gets the file paths and the simulation parameters.

startunit.f90:417

AquaCrop.initialize_transfer_assimilates! — Method

initialize_transfer_assimilates!(gvars, lvars)

run.f90:6378

AquaCrop.irri_manual! — Method

irri_manual!(gvars)

run.f90:6224

AquaCrop.irri_out_season — Method

irri = irri_out_season(gvars)

run.f90:6188

AquaCrop.isleapyear — Method

leapyear = isleapyear(year)

global.f90:7100

AquaCrop.ks_any — Method

ksval = ks_any(wrel, pulactual, pllactual, shapefactor)

global.f90:2611

AquaCrop.ks_salinity — Method

m = ks_salinity(salinityresponsconsidered, ecen, ecex, ecevar, ksshapesalinity)

global.f90:2043

AquaCrop.ks_temperature — Method

m = ks_temperature(t0, t1, tin)

global.f90:1981

AquaCrop.length_canopy_decline — Method

nd = length_canopy_decline(ccx, cdc)

global.f90:1839

AquaCrop.load_clim! — Method

load_clim!(record::RepClim, clim_file; kwargs...)

global.f90:5936

AquaCrop.load_crop! — Method

load_crop!(crop::RepCrop, perennial_period::RepPerennialPeriod, crop_file; kwargs...)

global.f90:4799

AquaCrop.load_groundwater! — Method

load_groundwater!(gvars, groundwater_file)

global.f90:5981

AquaCrop.load_gvars_from_toml! — Method

load_gvars_from_toml!(simulparam::RepParam, auxparfile; kwargs...)

AquaCrop.load_initial_conditions! — Method

 load_initial_conditions!(gvars, swcinifilefull)

global.f90:6486

AquaCrop.load_irri_schedule_info! — Method

load_irri_schedule_info!(gvars, irri_file)

global.f90:2860

AquaCrop.load_management! — Method

load_management!(gvars, man_file; kwargs...)

global.f90:3350

AquaCrop.load_offseason! — Method

load_offseason!(gvars, offseason_file)

global.f90:5769

AquaCrop.load_profile — Method

soil, soil_layers, compartments = load_profile(filepath, simulparam::RepParam; kwargs...)

loads data from filepath.

global.f90:7590

AquaCrop.load_profile_processing! — Method

load_profile_processing!(soil::RepSoil, soil_layers::Vector{SoilLayerIndividual},
    compartments::Vector{CompartmentIndividual}, simulparam::RepParam)

loads some data.

global.f90:7684

AquaCrop.load_program_parameters_project_plugin! — Method

load_program_parameters_project_plugin!(simulparam::RepParam, auxparfile; kwargs...)

startunit.f90:767

AquaCrop.load_simulation_project! — Method

load_simulation_project!(outputs, gvars, projectinput::ProjectInputType; kwargs...)

tempprocessing.f90:1932

AquaCrop.max_crat_depth — Method

crmax = max_crat_depth(paramcra, paramcrb, ksat, zi, depthgwt)

global.f90:2156

AquaCrop.mixing! — Method

mixing!(compartment::CompartmentIndividual, simulparam::RepParam, mm1, mm2, diff, ni)

simul.f90:2650

AquaCrop.move_salt_to! — Method

move_salt_to!(compx::CompartmentIndividual, simulparam::RepParam, soil_layers, celx_local, ds)

simul.f90:2675

AquaCrop.multiplier_cco_self_thinning — Method

fcco = multiplier_cco_self_thinning(yeari, yearx, shapefactor)

global.f90:2588

AquaCrop.multiplier_ccx_self_thinning — Method

fccx = multiplier_ccx_self_thinning(yeari, yearx, shapefactor)

global.f90:1784

AquaCrop.no_adjustment — Method

nadj = no_adjustment(fcvolpr)

global.f90:4195

AquaCrop.no_effect_stress! — Method

no_effect_stress!(theeffectstress::RepEffectStress)

simul.f90:4128

AquaCrop.no_irrigation! — Method

no_irrigation!(gvars)

global.f90:2838

AquaCrop.open_climfiles_and_get_data_firstday! — Method

open_climfiles_and_get_data_firstday!(outputs, gvars; kwargs...)

run.f90:5962

AquaCrop.open_harvest_info! — Method

open_harvest_info!(gvars, path; kwargs...)

run.f90:5937

AquaCrop.open_irrigation_file! — Method

open_irrigation_file!(gvars, path)

run.f90:4436

AquaCrop.prepare_stage1! — Method

prepare_stage1!(gvars)

simul.f90:4140

AquaCrop.prepare_stage2! — Method

prepare_stage2!(gvars)

simul.f90:4212

AquaCrop.read_climate_nextday! — Method

read_climate_nextday!(outputs, gvars)

run.f90:7345

AquaCrop.read_output_from_dayout — Method

read_output_from_dayout(outputs, i::Int)

AquaCrop.read_output_from_etodatasim — Method

tlow, thigh = read_output_from_etodatasim(outputs, i::Int)

AquaCrop.read_output_from_evaldataout — Method

read_output_from_evaldataout(outputs, i::Int)

AquaCrop.read_output_from_harvestsout — Method

read_output_from_harvestsout(outputs, i::Int)

AquaCrop.read_output_from_logger — Method

str = read_output_from_logger(outputs, i::Int)

AquaCrop.read_output_from_raindatasim — Method

tlow, thigh = read_output_from_raindatasim(outputs, i::Int)

AquaCrop.read_output_from_seasonout — Method

read_output_from_seasonout(outputs, i::Int)

AquaCrop.read_output_from_tcropsim — Method

tlow, thigh = read_output_from_tcropsim(outputs, i::Int)

AquaCrop.read_output_from_tempdatasim — Method

tlow, thigh = read_output_from_tempdatasim(outputs, i::Int)

AquaCrop.read_temperature_file! — Method

read_temperature_file!(array_parameters::ParametersContainer{T}, temperature_file) where T

tempprocessing.f90:307

AquaCrop.record_harvest! — Method

record_harvest!(outputs, gvars, nrcut, dayinseason, nrrun)

run.f90:6697

AquaCrop.relationships_for_fertility_and_salt_stress! — Method

relationships_for_fertility_and_salt_stress!(outputs, gvars)

run.f90:3954

AquaCrop.required_gdd — Method

requiredgdd = required_gdd(ccitofind, cco, ccx, gddcgcadjusted, gddayi, sumgddadjcc)

simul.f90:3841

AquaCrop.required_time_new — Method

t = required_time_new(ccitofind, cco, ccx, cgcadjusted, virtualtimecc)

simul.f90:5270

AquaCrop.reset_previous_sum! — Method

reset_previous_sum!(gvars)

run.f90:3445

AquaCrop.reset_swc_to_fc! — Method

reset_swc_to_fc!(simulation::RepSim, compartments::Vector{CompartmentIndividual},
                 soil_layers::Vector{SoilLayerIndividual}, ziaqua)

global.f90:4758

AquaCrop.root_max_in_soil_profile — Method

zmax = root_max_in_soil_profile(zmaxcrop, soil_layers::Vector{SoilLayerIndividual})

global.f90:1092

AquaCrop.run_simulation! — Method

run_simulation!(outputs, gvars, projectinput::Vector{ProjectInputType}; kwargs...)

run.f90:7779

AquaCrop.salt_solution_deposit! — Method

salt_solution_deposit!(compartment::CompartmentIndividual, simulparam::RepParam, i, mm)

global.f90:2572

AquaCrop.salt_transport_factor — Method

s = salt_transport_factor(theta, soil_layers)

simul.f90:4595

AquaCrop.seasonal_sum_of_kcpot — Method

sumkcpot = seasonal_sum_of_kcpot(outputs, thedaystoccini, thegddaystoccini, l0, l12, 
                                 l123, l1234, gddl0, gddl12, gddl123, 
                                 gddl1234, cco, ccx, cgc, gddcgc, cdc, 
                                 gddcdc, kctop, kcdeclageing, 
                                 cceffectprocent, tbase, tupper, tdaymin, 
                                 tdaymax, gdtransplow, co2i, themodecycle,
                                 simulation, simulparam)

global.f90:5315 note that we must do simulation.DelayedDays = 0 before calling this function

AquaCrop.set_clim_data! — Method

set_clim_data!(gvars, projectinput::ProjectInputType)

global.f90:4295

AquaCrop.set_daynr_to_yundef — Method

daynri = set_daynr_to_yundef(daynri)

tempprocessing.f90:351

AquaCrop.set_gdd_variables_nextday! — Method

set_gdd_variables_nextday!(gvars)

run.f90:7376

AquaCrop.setparameter! — Method

setparameter!(parameterscontainer::ParametersContainer{T}, parameterkey::Symbol, parameter::T)

sets an entry in parameterscontainer.

AquaCrop.soil_evaporation_reduction_coefficient — Method

s = soil_evaporation_reduction_coefficient(wrel, edecline)

global.f90:2139

AquaCrop.start_outputs — Method

outputs = start_outputs()

variable with the outputs of the program

AquaCrop.start_the_program — Function

start_the_program(parentdir::Union{String,Nothing}=nothing, runtype::Union{Symbol,Nothing}=nothing)

starts the program

startunit.f90:931

AquaCrop.stress_biomass_relationship! — Method

stress_biomass_relationship!(outputs, gvars)

tempprocessing.f90:3067

AquaCrop.sum_calendar_days — Method

nrcdays = sum_calendar_days(valgddays, firstdaycrop, tbase, tupper, tdaymin, tdaymax, gvars)

tempprocessing.f90:1035

AquaCrop.surface_transpiration! — Method

surface_transpiration!(gvars, coeffb0salt, coeffb1salt, coeffb2salt)

simul.f90:1523

AquaCrop.swcz_soil — Method

swcact = swcz_soil(gvars)

run.f90:6555

AquaCrop.tau_from_ksat — Method

tau = tau_from_ksat(ksat)

global.f90:1889

AquaCrop.temperature_file_covering_crop_period! — Method

temperature_file_covering_crop_period!(outputs, gvars, crop_firstday, crop_lastday)

tempprocessing.f90:1789

AquaCrop.time_root_function — Method

trf = time_root_function(t, shapefactor, tmax, t0)

global.f90:1263

AquaCrop.time_to_cc_ini — Method

elapsedtime = time_to_cc_ini(theplantingtype, thecropplantingdens, 
                      thesizeseedling, thesizeplant, thecropccx, thecropcgc)

global.f90:1754

AquaCrop.time_to_max_canopy_sf — Method

l12sf, redcgc, redccx, classsf = time_to_max_canopy_sf(cco, cgc, ccx, l0, l12, l123, ltoflor, lflor, determinantcrop, l12sf, redcgc, redccx, classsf)

global.f90:2090

AquaCrop.translate_inilayers_to_swprofile! — Method

translate_inilayers_to_swprofile!(gvars, nrlay, laythickness, layvolpr, layecds)

global.f90:6179

AquaCrop.translate_inipoints_to_swprofile! — Method

translate_inipoints_to_swprofile!(gvars, nrloc, locdepth, locvolpr, locecds)

global.f90:6258

AquaCrop.wc_evap_layer — Method

wx = wc_evap_layer(zlayer, attheta, compartments, soil_layers)

simul.f90:4158

AquaCrop.write_daily_results! — Method

write_daily_results!(outputs, gvars, dap, wpi, nrrun)

run.f90:7419

AquaCrop.write_evaluation_data! — Method

write_evaluation_data!(outputs, gvars, dap, nrrun)

run.f90:6498

AquaCrop.write_intermediate_period! — Method

write_intermediate_period!(outputs, gvars)

run.f90:6088

AquaCrop.write_sim_period! — Method

write_sim_period!(outputs, gvars, nrrun)

run.f90:6064

AquaCrop.write_the_results! — Method

write_the_results!(outputs, anumber, day1, month1, year1, dayn, monthn, 
                       yearn, rper, etoper, gddper, irriper, infiltper, 
                       roper, drainper, crwper, eper, exper, trper, trwper, 
                       trxper, salinper, saloutper, 
                       salcrper, biomassper, bunlimper, bmobper, bstoper, 
                       gvars)

run.f90:4582

AquaCrop.write_title_part1_mult_results! — Method

write_title_part1_mult_results!(outputs, gvars, nrrun)

run.f90:4545

AquaCrop.year_weighing_factor — Method

yeari = year_weighing_factor(cropfirstdaynr)

simul.f90:1073

AquaCrop.zr_adjusted_to_restrictive_layers — Method

zrout = zr_adjusted_to_restrictive_layers(zrin, soil_layers::Vector{SoilLayerIndividual})

global.f90:1126